requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
python-dotenv==1.0.0
bcrypt==4.0.1
transformers==4.33.3
# Torch version varies by Python/OS. Install a compatible build separately if needed.
torch>=2.6.0
numpy==1.26.4
pandas==2.1.0
python-dotenv==1.0.0
email-validator==2.0.0.post2
Flask-Migrate==4.0.5
python-dateutil==2.8.2
